+++ crates/serial-sensors-proto/src/data_types.rs
@@ -1,8 +1,10 @@
+mod identifier;
 mod scalar;
 mod vector2;
 mod vector3;
 mod vector4;
 
+pub use identifier::Identifier;
 pub use scalar::ScalarData;
 pub use vector2::Vector2Data;
 pub use vector3::Vector3Data;